+++ AddOns/TradeSkillMaster/ChangeLog.md
@@ -1,5 +1,6 @@
-## v4.11.64 Changes
+## v4.11.66 Changes
 
-* [Retail] Revert multi-crafting changes
+* [Retail] Fixed Destroy Next not working
+* [Retail] Fixed conflict with bag addons
 
 [Known Issues](https://support.tradeskillmaster.com/en_US/known_issues)